Moves Dual content into src
