+++ .gitignore
@@ -1,7 +1,4 @@
 *.bin
-frontend/vault-replica/.obsidian/plugins/Dual/node_modules/*
-frontend/vault-replica/.obsidian/plugins/Dual/coverage
-backend/target/*
-backend/models/*
-backend/index/*
-backend/emb_cache.db
+*.log
+node_modules
+coverage
