+++ package.json
@@ -22,12 +22,15 @@
     "@types/jest": "^26.0.23",
     "@types/node": "^14.14.2",
     "@types/remove-markdown": "^0.3.0",
+    "copyfiles": "^2.4.1",
     "i": "^0.3.6",
     "jest": "^26.6.3",
     "npm": "^7.19.0",
+    "npm-run-all": "^4.1.5",
     "obsidian": "^0.12.0",
     "prettier": "^2.3.0",
     "rollup": "^2.32.1",
+    "rollup-plugin-copy": "^3.4.0",
     "ts-jest": "^26.5.6",
     "tslib": "^2.0.3",
     "typescript": "^4.2.4"
